Moved index events route to directory
