+++ app/src/routes/Events/Index/components/EventsView.jsx
