+++ app/src/routes/Events/Index/components/filters/EventsFilter.scss
