+++ app/src/routes/Events/Index/components/filters/EventsFilters.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import pure from "recompose/pure";
 import Button from "react-mdl/lib/Button";
-import { Dialog, DialogTitle, DialogContent, DialogActions } from "../../../../components/Dialog";
+import { Dialog, DialogTitle, DialogContent, DialogActions } from "../../../../../components/Dialog";
 import LocationFilterContainer from "../../containers/LocationFilterContainer.jsx";
 import GamesFilterContainer from "../../containers/GamesFilterContainer.jsx";
 import TypeFilterContainer from "../../containers/TypeFilterContainer.jsx";
