+++ app/src/routes/Events/Index/components/filters/EventsFilters.scss
