+++ app/src/routes/Events/Index/components/filters/GamesFilter.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import pure from "recompose/pure";
-import Tags from "../../../../components/Tags";
+import Tags from "../../../../../components/Tags";
 import EventsFilter from "./EventsFilter.jsx";
 import "./GamesFilter.scss";
