+++ app/src/routes/Events/Index/components/filters/GamesFilter.scss
