+++ app/src/routes/Events/Index/components/filters/LocationFilter.scss
