+++ app/src/routes/Events/Index/components/filters/TypeFilter.scss
