+++ app/src/routes/Events/Index/components/list/EventsList.scss
