+++ app/src/routes/Events/Index/components/map/EventsMap.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import pure from "recompose/pure";
-import Map from "../../../../components/Map";
+import Map from "../../../../../components/Map";
 import MapPopup from "./MapPopup.jsx";
 import MapMarker from "./MapMarker.jsx";
 import MapAreaContainer from "../../containers/MapAreaContainer.jsx";
