+++ app/src/routes/Events/Index/components/map/EventsMap.scss
