+++ app/src/routes/Events/Index/components/map/MapArea.jsx
