+++ app/src/routes/Events/Index/components/map/MapMarker.jsx
