+++ app/src/routes/Events/Index/containers/MapAreaContainer.jsx
