+++ app/src/routes/Events/Index/modules/EventActions.js
@@ -1,4 +1,4 @@
-import Api from "../../../api";
+import Api from "../../../../api";
 import {
     SET_EVENTS
 } from "./EventConstants";
